feat: Implement initial data loader for banks
A system has been implemented to perform an initial data load when the application is started for the first time.

This ensures that the list of banks is available from the beginning. The logic has been centralized in an 'InitLoader' widget that is displayed on the splash screen and uses SharedPreferences to ensure that the load is only performed once.